requirements.txt
aiohttp==3.8.5
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
PyPDF2==3.0.1
reportlab==4.0.4
Werkzeug==2.3.7